AI-Agent/requirements.txt
strands-agents
strands-agents-tools

strands-agents[ollama]